There's audio, but it's glitched.
it only works at all with 128k or 256k non-psram builds

you can hear the beep when you adjust the slider in control panel, but
there's other audio overlaid on it that is kind of a weird descending tone,
very regular in nature.

The sample rate is set to 20k not 22.255k because of my i2s dac not
liking random sample rates.

Instead of using the i2s dac on the fruit jam, this uses A0 (data),
A1 (lrck), A2 (bclk). It's not super convenient, and it's hard coded.